+++ R/data.R
@@ -134,6 +134,7 @@
 #' \item{e_dot_perc}{The name of the energy flow percentage column in final-to-useful templates.}
 #' \item{e_dot_machine}{The name of the column representing energy flow into a machine in final-to-useful templates.}
 #' \item{e_dot_machine_perc}{The name of the machine energy flow percentage column in final-to-useful templates.}
+#' \item{e_dot_machine_max_perc}{The name of the machine energy flow maximum percentage column in final-to-useful templates.}
 #' \item{eta_fu}{The name of the final-to-useful machine efficiency column in final-to-useful templates.}
 #' \item{phi_u}{The name of the exergy-to-energy ratio column in final-to-useful templates.}
 #' \item{c_source}{The name of a column containing country identifiers for the source of final-to-useful allocation information.}